refactor: 이력서 관리 - 마이페이지로 이동
