--- src/pages/ResumePages/ResumeManagementPage/index.ts
@@ -1,3 +1,0 @@
-import ResumeManagementPage from './ResumeManagementPage';
-
-export { ResumeManagementPage };
